Split some of PackageFileService and FileStorageService into NuGetGallery.Core (#4759)
Progress on https://github.com/NuGet/NuGetGallery/issues/4646
