+++ src/NuGetGallery.Core/NuGetGallery.Core.csproj
@@ -205,8 +205,18 @@
     <Compile Include="Properties\AssemblyInfo.*.cs" />
     <Compile Include="NuGetVersionExtensions.cs" />
     <Compile Include="SemVerLevelKey.cs" />
+    <Compile Include="Services\CloudBlobCoreFileStorageService.cs" />
+    <Compile Include="Services\CloudFileReference.cs" />
+    <Compile Include="Services\CorePackageFileService.cs" />
     <Compile Include="Services\CorePackageService.cs" />
+    <Compile Include="Services\ICloudBlobClient.cs" />
+    <Compile Include="Services\ICloudBlobContainer.cs" />
+    <Compile Include="Services\ICorePackageFileService.cs" />
     <Compile Include="Services\ICorePackageService.cs" />
+    <Compile Include="Services\IFileReference.cs" />
+    <Compile Include="Services\ICoreFileStorageService.cs" />
+    <Compile Include="Services\ISimpleCloudBlob.cs" />
+    <Compile Include="Services\TestableStorageClientException.cs" />
     <Compile Include="StreamExtensions.cs" />
     <Compile Include="CoreStrings.Designer.cs">
       <AutoGen>True</AutoGen>
