+++ src/NuGetGallery.Core/Services/CloudFileReference.cs
