+++ src/NuGetGallery.Core/Services/ICloudBlobClient.cs
