+++ src/NuGetGallery.Core/Services/ICloudBlobContainer.cs
