+++ src/NuGetGallery.Core/Services/IFileReference.cs
