+++ src/NuGetGallery.Core/Services/ISimpleCloudBlob.cs
