+++ src/NuGetGallery.Core/Services/TestableStorageClientException.cs
