+++ src/NuGetGallery/Constants.cs
@@ -33,24 +33,13 @@ namespace NuGetGallery
 
         public const int MaxEmailSubjectLength = 255;
         internal static readonly NuGetVersion MaxSupportedMinClientVersion = new NuGetVersion("4.1.0.0");
-        public const string PackageContentType = "binary/octet-stream";
-        public const string OctetStreamContentType = "application/octet-stream";
-        public const string TextContentType = "text/plain";
         public const string PackageFileDownloadUriTemplate = "packages/{0}/{1}/download";
-        public const string PackageFileSavePathTemplate = "{0}.{1}{2}";
 
         public const string ReadMeFileSavePathTemplateActive = "active/{0}/{1}{2}";
         public const string ReadMeFileSavePathTemplatePending = "pending/{0}/{1}{2}";
 
         public const string PackageFileBackupSavePathTemplate = "{0}/{1}/{2}.{3}";
 
-        public const string PackagesFolderName = "packages";
-        public const string PackageBackupsFolderName = "package-backups";
-        public const string DownloadsFolderName = "downloads";
-        public const string ContentFolderName = "content";
-        public const string PackageReadMesFolderName = "readmes";
-
-        public const string NuGetPackageFileExtension = ".nupkg";
         public const string MarkdownFileExtension = ".md";
         public const string HtmlFileExtension = ".html";
         public const string JsonFileExtension = ".json";
@@ -64,7 +53,6 @@ namespace NuGetGallery
         public const string PBKDF2HashAlgorithmId = "PBKDF2";
 
         public const string UploadFileNameTemplate = "{0}{1}";
-        public const string UploadsFolderName = "uploads";
         public const string NuGetCommandLinePackageId = "NuGet.CommandLine";
 
         public static readonly string ReturnUrlViewDataKey = "ReturnUrl";
