+++ src/NuGetGallery/Services/PackageFileService.cs
@@ -12,7 +12,7 @@ using NuGet.Versioning;
 
 namespace NuGetGallery
 {
-    public class PackageFileService : IPackageFileService
+    public class PackageFileService : CorePackageFileService, IPackageFileService
     {
         /// <summary>
         /// Active readme markdown file, formatted as 'active/{packageId}/{version}.md'
@@ -27,20 +27,21 @@ namespace NuGetGallery
         private readonly IFileStorageService _fileStorageService;
 
         public PackageFileService(IFileStorageService fileStorageService)
+            : base(fileStorageService)
         {
             _fileStorageService = fileStorageService;
         }
 
         public Task<ActionResult> CreateDownloadPackageActionResultAsync(Uri requestUrl, Package package)
         {
-            var fileName = BuildFileName(package, Constants.PackageFileSavePathTemplate, Constants.NuGetPackageFileExtension);
-            return _fileStorageService.CreateDownloadFileActionResultAsync(requestUrl, Constants.PackagesFolderName, fileName);
+            var fileName = BuildFileName(package, CoreConstants.PackageFileSavePathTemplate, CoreConstants.NuGetPackageFileExtension);
+            return _fileStorageService.CreateDownloadFileActionResultAsync(requestUrl, CoreConstants.PackagesFolderName, fileName);
         }
 
         public Task<ActionResult> CreateDownloadPackageActionResultAsync(Uri requestUrl, string id, string version)
         {
-            var fileName = BuildFileName(id, version, Constants.PackageFileSavePathTemplate, Constants.NuGetPackageFileExtension);
-            return _fileStorageService.CreateDownloadFileActionResultAsync(requestUrl, Constants.PackagesFolderName, fileName);
+            var fileName = BuildFileName(id, version, CoreConstants.PackageFileSavePathTemplate, CoreConstants.NuGetPackageFileExtension);
+            return _fileStorageService.CreateDownloadFileActionResultAsync(requestUrl, CoreConstants.PackagesFolderName, fileName);
         }
 
         public Task DeletePackageFileAsync(string id, string version)
@@ -55,19 +56,8 @@ namespace NuGetGallery
                 throw new ArgumentNullException(nameof(version));
             }
 
-            var fileName = BuildFileName(id, version, Constants.PackageFileSavePathTemplate, Constants.NuGetPackageFileExtension);
-            return _fileStorageService.DeleteFileAsync(Constants.PackagesFolderName, fileName);
-        }
-
-        public Task SavePackageFileAsync(Package package, Stream packageFile)
-        {
-            if (packageFile == null)
-            {
-                throw new ArgumentNullException(nameof(packageFile));
-            }
-
-            var fileName = BuildFileName(package, Constants.PackageFileSavePathTemplate, Constants.NuGetPackageFileExtension);
-            return _fileStorageService.SaveFileAsync(Constants.PackagesFolderName, fileName, packageFile, overwrite: false);
+            var fileName = BuildFileName(id, version, CoreConstants.PackageFileSavePathTemplate, CoreConstants.NuGetPackageFileExtension);
+            return _fileStorageService.DeleteFileAsync(CoreConstants.PackagesFolderName, fileName);
         }
 
         public Task StorePackageFileInBackupLocationAsync(Package package, Stream packageFile)
@@ -86,18 +76,12 @@ namespace NuGetGallery
                 string.IsNullOrWhiteSpace(package.PackageRegistration.Id) ||
                 (string.IsNullOrWhiteSpace(package.NormalizedVersion) && string.IsNullOrWhiteSpace(package.Version)))
             {
-                throw new ArgumentException(Strings.PackageIsMissingRequiredData, nameof(package));
+                throw new ArgumentException(CoreStrings.PackageIsMissingRequiredData, nameof(package));
             }
 
             var fileName = BuildBackupFileName(package.PackageRegistration.Id, string.IsNullOrEmpty(package.NormalizedVersion) 
                 ? NuGetVersion.Parse(package.Version).ToNormalizedString() : package.NormalizedVersion, package.Hash);
-            return _fileStorageService.SaveFileAsync(Constants.PackageBackupsFolderName, fileName, packageFile);
-        }
-
-        public Task<Stream> DownloadPackageFileAsync(Package package)
-        {
-            var fileName = BuildFileName(package, Constants.PackageFileSavePathTemplate, Constants.NuGetPackageFileExtension);
-            return _fileStorageService.GetFileAsync(Constants.PackagesFolderName, fileName);
+            return _fileStorageService.SaveFileAsync(CoreConstants.PackageBackupsFolderName, fileName, packageFile);
         }
 
         /// <summary>
@@ -117,7 +101,7 @@ namespace NuGetGallery
                 ReadMeFilePathTemplateActive;
             var fileName = BuildFileName(package, format, Constants.MarkdownFileExtension);
 
-            return _fileStorageService.DeleteFileAsync(Constants.PackageReadMesFolderName, fileName);
+            return _fileStorageService.DeleteFileAsync(CoreConstants.PackageReadMesFolderName, fileName);
         }
 
         /// <summary>
@@ -135,7 +119,7 @@ namespace NuGetGallery
             var fileName = BuildFileName(package, ReadMeFilePathTemplatePending, Constants.MarkdownFileExtension);
             using (var readMeMdStream = new MemoryStream(Encoding.UTF8.GetBytes(readMeMd)))
             {
-                await _fileStorageService.SaveFileAsync(Constants.PackageReadMesFolderName, fileName, readMeMdStream, overwrite: true);
+                await _fileStorageService.SaveFileAsync(CoreConstants.PackageReadMesFolderName, fileName, readMeMdStream, overwrite: true);
             }
         }
 
@@ -156,7 +140,7 @@ namespace NuGetGallery
                 ReadMeFilePathTemplateActive;
             var fileName = BuildFileName(package, format, Constants.MarkdownFileExtension);
 
-            using (var readMeMdStream = await _fileStorageService.GetFileAsync(Constants.PackageReadMesFolderName, fileName))
+            using (var readMeMdStream = await _fileStorageService.GetFileAsync(CoreConstants.PackageReadMesFolderName, fileName))
             {
                 // Note that fileStorageService implementations return null if not found.
                 if (readMeMdStream != null)
@@ -169,52 +153,6 @@ namespace NuGetGallery
             }
 
             return null;
-        }
-
-        private static string BuildFileName(string id, string version, string pathTemplate, string extension)
-        {
-            if (id == null)
-            {
-                throw new ArgumentNullException(nameof(id));
-            }
-
-            if (version == null)
-            {
-                throw new ArgumentNullException(nameof(version));
-            }
-
-            // Note: packages should be saved and retrieved in blob storage using the lower case version of their filename because
-            // a) package IDs can and did change case over time
-            // b) blob storage is case sensitive
-            // c) we don't want to hit the database just to look up the right case
-            // and remember - version can contain letters too.
-            return String.Format(
-                CultureInfo.InvariantCulture,
-                pathTemplate,
-                id.ToLowerInvariant(),
-                version.ToLowerInvariant(),
-                extension);
-        }
-
-        private static string BuildFileName(Package package, string format, string extension)
-        {
-            if (package == null)
-            {
-                throw new ArgumentNullException(nameof(package));
-            }
-
-            if (package.PackageRegistration == null ||
-                String.IsNullOrWhiteSpace(package.PackageRegistration.Id) ||
-                (String.IsNullOrWhiteSpace(package.NormalizedVersion) && String.IsNullOrWhiteSpace(package.Version)))
-            {
-                throw new ArgumentException(Strings.PackageIsMissingRequiredData, nameof(package));
-            }
-
-            return BuildFileName(
-                package.PackageRegistration.Id,
-                string.IsNullOrEmpty(package.NormalizedVersion) ?
-                    NuGetVersionFormatter.Normalize(package.Version) :
-                    package.NormalizedVersion, format, extension);
         }
 
         private static string BuildBackupFileName(string id, string version, string hash)
@@ -242,7 +180,7 @@ namespace NuGetGallery
                 id,
                 version,
                 HttpServerUtility.UrlTokenEncode(hashBytes),
-                Constants.NuGetPackageFileExtension);
+                CoreConstants.NuGetPackageFileExtension);
         }
     }
 }
