+++ src/NuGetGallery/Strings.resx
@@ -320,9 +320,6 @@ The {1} Team</value>
   <data name="JobLogBlobNameInvalid" xml:space="preserve">
     <value>Job Log Blob name is invalid Bob! Expected [jobname].[yyyy-MM-dd].json or [jobname].json. Got: {0}</value>
   </data>
-  <data name="PackageIsMissingRequiredData" xml:space="preserve">
-    <value>The package is missing required data.</value>
-  </data>
   <data name="ErrorInSendingMail" xml:space="preserve">
     <value>Error in sending mail : {0}</value>
   </data>
